requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
jax>=0.4.1
prefect>=2.10.0
pandas>=1.5.0
numpy>=1.23.0
plotly==5.18.0
tenacity>=8.2.3
dash>=2.9.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
scikit-learn>=1.2.0
nltk>=3.8.1
kubernetes>=26.1.0
psutil>=5.9.0
pyyaml>=6.0.0
gputil>=1.4.0 